+++ core/farneback_optical_flow_processor.cc
@@ -81,8 +81,9 @@ FarnebackOpticalFlowProcessor::Process(
   cv::Mat first_gray = first, second_gray = second, flow;
 
   cv::calcOpticalFlowFarneback(first_gray, second_gray, flow,
-                               pyr_scale_, levels_, winsize_, iterations_,
-                               poly_n_, poly_sigma_, flags_);
+                               pyr_scale_, (int) levels_, (int) winsize_, 
+                               (int) iterations_, (int) poly_n_, poly_sigma_, 
+                               (int) flags_);
 
   std::vector< std::pair<cv::Point2d, cv::Point2d> > v;
